+++ comics/comics/reveland.py
@@ -7,15 +7,11 @@ class ComicData(ComicDataBase):
     language = 'no'
     url = 'http://reveland.nettserier.no/'
     start_date = '2007-03-20'
+    end_date = '2013-04-17'
+    active = False
     rights = 'Jorunn Hanto-Haugse'
 
 
 class Crawler(CrawlerBase):
-    history_capable_days = 90
-    schedule = 'Mo,We,Fr'
-    time_zone = 'Europe/Oslo'
-
     def crawl(self, pub_date):
-        epoch = self.date_to_epoch(pub_date)
-        url = 'http://reveland.nettserier.no/_striper/reveland-%s.jpg' % epoch
-        return CrawlerImage(url)
+        pass  # Comic no longer published